MAINT: stats: separate warnings and errors from result classes
