+++ doc/source/reference/stats._warnings_errors.rst
@@ -1,0 +1,4 @@
+.. automodule:: scipy.stats._warnings_errors
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
