+++ scipy/stats/__init__.py
@@ -449,6 +449,15 @@ Univariate and multivariate kernel density estimation
 
    gaussian_kde
 
+Warnings / Errors used in :mod:`scipy.stats`
+--------------------------------------------
+
+.. autosummary::
+   :toctree: generated/
+
+   SciPyStatsError
+   SciPyStatsWarning
+
 """
 
 from ._stats_py import *
@@ -470,6 +479,7 @@ from ._rvs_sampling import rvs_ratio_uniforms, NumericalInverseHermite  # noqa
 from ._page_trend_test import page_trend_test
 from ._mannwhitneyu import mannwhitneyu
 from ._fit import fit
+from ._warnings_errors import SciPyStatsError, SciPyStatsWarning
 
 # Deprecated namespaces, to be removed in v2.0.0
 from . import (
