+++ scipy/stats/meson.build
@@ -214,6 +214,7 @@ py3.install_sources([
     '_stats_py.py',
     '_tukeylambda_stats.py',
     '_variation.py',
+    '_warnings_errors.py',
     'biasedurn.py',
     'contingency.py',
     'distributions.py',
